ci(GHA): Resolve commitlint failures
This makes the 'check commits' GitHub Actions job use a pre-authenticated GitHub API client by using the actions/github-script@v6 action.

This should resolve rate limit errors which were previously happening.

These are the same changes that were made in https://github.com/defencedigital/mod-uk-design-system/pull/3566.
