+++ .github/workflows/build_and_test.yml
@@ -21,11 +21,18 @@ jobs:
             path: '**/node_modules'
             key: ${{ runner.os }}-modules-${{ hashFiles('**/yarn.lock') }}
 
-        - name: check commits
-          run: |
-            yarn install --cache-folder ~/.cache/yarn
-            node ./scripts/commitlint ${{ github.event.pull_request._links.html.href }}
-            node ./scripts/check-fixup ${{ github.event.pull_request._links.html.href }}
+        - name: Install dependencies
+          run: yarn install --cache-folder ~/.cache/yarn
+
+        - name: Check commits
+          uses: actions/github-script@v6
+          with:
+            script: |
+              const { runCommitLint } = await import('${{ github.workspace }}/scripts/github-actions/runCommitlint.mjs')
+              await runCommitLint({ github, context, core })
+
+              const { checkForFixups } = await import('${{ github.workspace }}/scripts/github-actions/checkForFixups.mjs')
+              await checkForFixups({ github, context, core })
 
     Security_audit:
       runs-on: ubuntu-latest
